raop: Move packets to support
Generalize the "packets" module and rename it to just packet.

Relates to #1255
